refactor(gateway-client): own the channel envelope so it carries the canonical verdict (LUM-2928)
Addresses the three blockers on #39593.

1. The envelope moves up into `@vellumai/gateway-client`, option (a). The
   authority block has to carry the real `TrustVerdictSchema` and
   `AdmissionPolicySchema`, and both already live here; `service-contracts` sits
   below this package, so an envelope defined down there could only reference a
   copy, and the copy I shipped was a 3-field strict subset that rejected a real
   guardian verdict outright (`unrecognized_keys` on guardianExternalUserId,
   guardianPrincipalId, guardianDisplayName, contactId, interactionCount).
   Composing upward gives one definition of each instead of three. It is also
   what `gateway/AGENTS.md` already says this package is for: schemas that cross
   the gateway/daemon/web wire.

   `service-contracts` keeps the provider vocabulary that knows nothing about
   trust: address, actor, capabilities, and the event-kind and content-ref
   primitives (now in `channel-event.ts`). The duplicated TRUST_CLASS_VALUES and
   ADMISSION_POLICY_VALUES are gone rather than deferred.

2. `channelEnvelopeV1()` extended the unrefined field set, so the payload
   path silently skipped the address/actor same-channel check while
   `ChannelEnvelopeV1ConsistentSchema` enforced it: the ergonomic path was the
   unsafe one. The field set is now private, the refinement is a shared
   function, and both the plain schema and the payload helper apply it, so there
   is no public unrefined path left to reach for. Regression tests assert
   rejection through the helper specifically, not only the plain schema.

3. The `stampedBy` docstring claimed an envelope assembled elsewhere fails
   validation. It does not: any assembler writes the literal. Reworded to say
   what the constant does (catches a malformed assembler, documents who fills
   the block) and to state plainly that authority rests on the transport, with
   the daemon not being publicly reachable as the actual boundary, and that
   accepting envelopes from an untrusted party would need a signature.
