+++ packages/service-contracts/package.json
@@ -11,7 +11,7 @@
     "./channel-address": "./src/channel-address.ts",
     "./channel-capabilities": "./src/channel-capabilities.ts",
     "./channel-coordinates": "./src/channel-coordinates.ts",
-    "./channel-envelope": "./src/channel-envelope.ts",
+    "./channel-event": "./src/channel-event.ts",
     "./client-metadata": "./src/client-metadata.ts",
     "./credential-rpc": "./src/credential-rpc.ts",
     "./ingress": "./src/ingress.ts",
